Revert unrelated test changes.
