+++ extensions/amp-analytics/0.1/iframe-transport-vendors.js
@@ -24,6 +24,6 @@
  *
  * @const {!JsonObject}
  */
-export const IFRAME_TRANSPORTS = {
+export const IFRAME_TRANSPORTS = /** @type {!JsonObject} */ ({
   'bg': 'https://tpc.googlesyndication.com/b4a/b4a-runner.html',
-};
+});
